feat.: implement jwt authentication and authorization

+++ src/main/java/cat/linky/linky_cat_api/core/service/link/LinkDeleteService.java
@@ -2,6 +2,7 @@ package cat.linky.linky_cat_api.core.service.link;
 
 import java.util.UUID;
 
+import cat.linky.linky_cat_api.core.exception.UnauthorizedOperationException;
 import cat.linky.linky_cat_api.core.ports.in.usecase.link.LinkDeleteUseCase;
 import cat.linky.linky_cat_api.core.ports.out.repository.LinkRepositoryPort;
 
@@ -14,7 +15,10 @@ public class LinkDeleteService implements LinkDeleteUseCase {
     }
 
     @Override
-    public void execute(UUID id) {
+    public void execute(UUID id, UUID userId) {
+        if (!repositoryPort.checkOwnership(id, userId))
+            throw new UnauthorizedOperationException();
+
         repositoryPort.deleteById(id);
     }
 }
